Split calculation in GetMatrixElement(LevelVector) into two separate loops over small and big configs.
There is no way to conditionally generate a task in OpenMP (the ``if'' clause only determines whether to
defer the task, not whether to generate it with all the overhead that entails), so it's necessary to
split the calculation into two loops. Configurations with more than some fraction of the total
projections (currently 1%) are deferred until a second loop, which then gets parallelised with one task
per pair of projections.