Fixed another bug when reading the Hamiltonian matrix from file.
When we're reading from scratch (i.e. instead of generating the matrix),
then each rank starts with no idea which chunks it "owns" (as this is
calculated in the body of the GenerateMatrix() function). We therefore
need to add in some logic to calculate the distribution of matrix chunks
when reading from file.