Add a basic continuous integration workflow
Adds a GitHub action configuration to automatically build and run unit tests (on a GitHub-hosted Ubuntu VM (currently free because we're public and open-source)) whenever there are pushes or pull requests to master and dev branches. This means that whenever we make changes, it will automatically spin off the test suite and alert us if anything is broken.

Currently, it doesn't do anything other than alert, but eventually it would be good to block any PRs with failing tests until they are fixed, for an extra level of safety. 

It also currently only runs in a clean Ubuntu VM, so it's not a perfect fit for production environments. However, if this action proves to be useful then we can see about using other, self-hosted runners which more closely match our production environments (e.g. HPC clusters).